re-organise project constructure, add folder client and server
