+++ src/client/components/Content.js
