+++ src/client/components/ContentEditor.js
